a_l_l_e_g_r_o_pl/Игорь_@swaytornado_monitoring/requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
et_xmlfile==2.0.0
idna==3.10
loguru==0.7.2
lxml==5.3.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.5.0
six==1.16.0
soupsieve==2.6
tqdm==4.67.0
tzdata==2024.2
urllib3==2.2.3
win32-setctime==1.1.0
